# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Panacea Socios API",
    "description": "API para gesti\u00f3n de socios, sus preferencias y registro de ventas de Panacea. Permite registrar clientes, capturar sus respuestas a preguntas clave y llevar un historial de compras para futuras promociones y regalos.",
    "contact": {
      "name": "Panacea",
      "email": "info@panacea.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/ventas/": {
      "get": {
        "tags": [
          "Registro de Ventas"
        ],
        "summary": "Listar ventas",
        "description": "Retorna ventas con sus detalles. Filtrable por socio_id. Requiere admin.",
        "operationId": "list_ventas_ventas__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "socio_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Socio Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VentaResponse"
                  },
                  "title": "Response List Ventas Ventas  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Registro de Ventas"
        ],
        "summary": "Registrar venta",
        "description": "Crea un registro de venta con sus l\u00edneas de detalle. Requiere admin.",
        "operationId": "create_venta_ventas__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VentaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ventas/{venta_id}": {
      "get": {
        "tags": [
          "Registro de Ventas"
        ],
        "summary": "Obtener venta",
        "description": "Retorna una venta con sus detalles por ID. Requiere admin.",
        "operationId": "get_venta_ventas__venta_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "venta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Venta Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Registro de Ventas"
        ],
        "summary": "Actualizar cabecera de venta",
        "description": "Actualiza los campos de cabecera de una venta. Requiere admin.",
        "operationId": "update_venta_ventas__venta_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "venta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Venta Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VentaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Registro de Ventas"
        ],
        "summary": "Eliminar venta",
        "description": "Elimina una venta y todos sus detalles (cascade). Requiere admin.",
        "operationId": "delete_venta_ventas__venta_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "venta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Venta Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DetalleCreate": {
        "properties": {
          "producto_id": {
            "type": "string",
            "maxLength": 100,
            "title": "Producto Id",
            "examples": [
              "CERVEZA-IPA-500ML"
            ]
          },
          "cantidad": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Cantidad",
            "examples": [
              2
            ]
          },
          "subtotal": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Subtotal",
            "examples": [
              [
                "350.00"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "producto_id",
          "cantidad",
          "subtotal"
        ],
        "title": "DetalleCreate"
      },
      "DetalleResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "registro_venta_id": {
            "type": "integer",
            "title": "Registro Venta Id"
          },
          "producto_id": {
            "type": "string",
            "title": "Producto Id"
          },
          "cantidad": {
            "type": "integer",
            "title": "Cantidad"
          },
          "subtotal": {
            "type": "string",
            "title": "Subtotal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "registro_venta_id",
          "producto_id",
          "cantidad",
          "subtotal"
        ],
        "title": "DetalleResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VentaCreate": {
        "properties": {
          "socio_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Socio Id",
            "examples": [
              1
            ]
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha",
            "examples": [
              "2024-06-15T20:30:00"
            ]
          },
          "lugar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Lugar",
            "examples": [
              "Local Palermo"
            ]
          },
          "importe": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Importe",
            "examples": [
              [
                "700.00"
              ]
            ]
          },
          "detalles": {
            "items": {
              "$ref": "#/components/schemas/DetalleCreate"
            },
            "type": "array",
            "title": "Detalles"
          }
        },
        "type": "object",
        "required": [
          "fecha",
          "importe"
        ],
        "title": "VentaCreate"
      },
      "VentaResponse": {
        "properties": {
          "socio_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Socio Id",
            "examples": [
              1
            ]
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha",
            "examples": [
              "2024-06-15T20:30:00"
            ]
          },
          "lugar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Lugar",
            "examples": [
              "Local Palermo"
            ]
          },
          "importe": {
            "type": "string",
            "title": "Importe",
            "examples": [
              [
                "700.00"
              ]
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "detalles": {
            "items": {
              "$ref": "#/components/schemas/DetalleResponse"
            },
            "type": "array",
            "title": "Detalles",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "fecha",
          "importe",
          "id"
        ],
        "title": "VentaResponse"
      },
      "VentaUpdate": {
        "properties": {
          "socio_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Socio Id"
          },
          "fecha": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha"
          },
          "lugar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Lugar"
          },
          "importe": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importe"
          }
        },
        "type": "object",
        "title": "VentaUpdate"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}